fix: Have clear init state with page available
