+++ packages/midi-bricks/src/components/drawer-list/DrawerList.jsx
@@ -18,7 +18,7 @@ import ViewSettingsDialog from '../ApplicationSettingsDialog'
 import { ListItemLoadFileOnElectron } from './ListItemLoadFileOnElectron'
 import { ListItemSaveFileOnElectron } from './ListItemSaveFileOnElectron'
 import { ListItemLoadFileOnWeb } from './ListItemLoadFileOnWeb'
-import { PAGE_TYPES } from '../../reducers/view-settings'
+import { PAGE_TYPES } from '../../reducers'
 import { PropTypes } from 'prop-types'
 
 DrawerList.propTypes = {
