+++ packages/midi-bricks/src/components/midi-settings/elements/DriverEmtpyRedirectButton.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Button from '@material-ui/core/Button'
 import Tooltip from '@material-ui/core/Tooltip'
-import { PAGE_TYPES } from '../../../reducers/view-settings'
+import { PAGE_TYPES } from '../../../reducers'
 
 export function DriverEmtpyRedirectButton({
   i,
